bugs fixed

+++ Services/VaultsService.cs
@@ -41,7 +41,7 @@ namespace Keepr.Services
     {
       Vault original = _repo.GetById(id);
       if (original == null) { throw new Exception("Invalid Id"); }
-      // if (original.CreatorId != creatorId) { throw new Exception("Access Denied... This is not yours"); }
+      if (original.CreatorId != creatorId) { throw new Exception("Access Denied... This is not yours"); }
       _repo.Remove(id);
       return "succesfully delorted";
